chore(events): factor out ‹EventData›
