+++ packit_service/worker/allowlist.py
@@ -34,7 +34,7 @@ from packit_service.worker.events import (
     pagure,
     testing_farm,
 )
-from packit_service.worker.events.event import EventData
+from packit_service.worker.events.event_data import EventData
 from packit_service.worker.helpers.build import CoprBuildJobHelper
 from packit_service.worker.helpers.testing_farm import TestingFarmJobHelper
 from packit_service.worker.reporting import BaseCommitStatus
