+++ packit_service/worker/handlers/mixin.py
@@ -28,7 +28,7 @@ from packit_service.worker.events import (
     koji,
     pagure,
 )
-from packit_service.worker.events.event import EventData
+from packit_service.worker.events.event_data import EventData
 from packit_service.worker.handlers.abstract import CeleryTask
 from packit_service.worker.helpers.build.copr_build import CoprBuildJobHelper
 from packit_service.worker.helpers.build.koji_build import KojiBuildJobHelper
