+++ packit_service/worker/helpers/fedora_ci.py
@@ -5,7 +5,7 @@ import logging
 
 from ogr.abstract import GitProject
 
-from packit_service.worker.events.event import EventData
+from packit_service.worker.events.event_data import EventData
 from packit_service.worker.reporting import BaseCommitStatus
 from packit_service.worker.reporting.reporters.base import StatusReporter
